5~10% faster unsafe compressor
